% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_find_clusters.R
\name{cifti_find_clusters}
\alias{cifti_find_clusters}
\title{Find clusters in a CIFTI file}
\usage{
cifti_find_clusters(
  cifti_in,
  surface_value_threshold,
  surface_minimum_area,
  volume_value_threshold,
  volume_minimum_size,
  direction = c("COLUMN", "ROW"),
  cifti_out,
  left_surface = NULL,
  right_surface = NULL,
  less_than = FALSE,
  verbose = get_wb_verbosity()
)
}
\arguments{
\item{cifti_in}{Path to the input CIFTI file.}

\item{surface_value_threshold}{Minimum value for surface vertices.}

\item{surface_minimum_area}{Minimum cluster surface area (mm^2).}

\item{volume_value_threshold}{Minimum value for volume voxels.}

\item{volume_minimum_size}{Minimum cluster volume size (mm^3).}

\item{direction}{Direction to use. One of \code{"ROW"} or \code{"COLUMN"}.}

\item{cifti_out}{Path for the output CIFTI file.}

\item{left_surface}{Path to the left hemisphere surface.}

\item{right_surface}{Path to the right hemisphere surface.}

\item{less_than}{Logical; find clusters below threshold instead of above?}

\item{verbose}{Logical; print command output?}
}
\value{
The result of the underlying \code{\link[=wb_cmd]{wb_cmd()}} call.
}
\description{
Wraps \code{wb_command -cifti-find-clusters} to threshold a CIFTI file and
filter clusters by surface area and volume size.
}
\examples{
\dontrun{
cifti_find_clusters(
  "zstats.dscalar.nii",
  surface_value_threshold = 2.3,
  surface_minimum_area = 100,
  volume_value_threshold = 2.3,
  volume_minimum_size = 200,
  direction = "COLUMN",
  cifti_out = "clusters.dscalar.nii",
  left_surface = "left.midthickness.surf.gii",
  right_surface = "right.midthickness.surf.gii"
)
}

}
\section{Connectome Workbench Help}{
\Sexpr[results=rd,stage=render]{wbcmd::wb_help_rd("-cifti-find-clusters")}
}

