% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_reduce.R
\name{cifti_reduce}
\alias{cifti_reduce}
\title{Reduce a CIFTI file along a dimension}
\usage{
cifti_reduce(
  cifti_in,
  direction = c("ROW", "COLUMN"),
  operation = "MEAN",
  cifti_out,
  only_numeric = FALSE,
  verbose = get_wb_verbosity()
)
}
\arguments{
\item{cifti_in}{Path to the input CIFTI file.}

\item{direction}{Direction to reduce. One of \code{"ROW"} or \code{"COLUMN"}.}

\item{operation}{Reduction operation. One of \code{"MAX"}, \code{"MIN"}, \code{"INDEXMAX"},
\code{"INDEXMIN"}, \code{"SUM"}, \code{"PRODUCT"}, \code{"MEAN"}, \code{"STDEV"}, \code{"SAMPSTDEV"},
\code{"VARIANCE"}, \code{"TSNR"}, \code{"COV"}, \code{"L2NORM"}, \code{"MEDIAN"}, \code{"MODE"},
\code{"COUNT_NONZERO"}.}

\item{cifti_out}{Path for the output CIFTI file.}

\item{only_numeric}{Logical; exclude non-numeric values?}

\item{verbose}{Logical; print command output?}
}
\value{
The result of the underlying \code{\link[=wb_cmd]{wb_cmd()}} call.
}
\description{
Wraps \code{wb_command -cifti-reduce} to perform a reduction operation
(e.g. mean, stdev) along a specified direction.
}
\examples{
\dontrun{
cifti_reduce(
  "data.dtseries.nii",
  direction = "ROW",
  operation = "MEAN",
  cifti_out = "mean.dscalar.nii"
)
}

}
