% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_information.R
\name{file_information}
\alias{file_information}
\title{Display information about a file}
\usage{
file_information(
  data_file,
  only_map_names = FALSE,
  only_number_of_maps = FALSE,
  only_metadata = FALSE,
  only_cifti_xml = FALSE,
  verbose = get_wb_verbosity()
)
}
\arguments{
\item{data_file}{Path to the input file.}

\item{only_map_names}{Logical; show only map names?}

\item{only_number_of_maps}{Logical; show only number of maps?}

\item{only_metadata}{Logical; show only metadata?}

\item{only_cifti_xml}{Logical; show only CIFTI XML (CIFTI files only)?}

\item{verbose}{Logical; print command output?}
}
\value{
Character vector of information lines, invisibly.
}
\description{
Wraps \code{wb_command -file-information} to list information about
a data file's content.
}
\examples{
\dontrun{
file_information("data.dscalar.nii")
file_information("data.dscalar.nii", only_map_names = TRUE)
}

}
