% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wb_setting.R
\name{get_wb_verbosity}
\alias{get_wb_verbosity}
\title{Get the workbench verbosity setting}
\usage{
get_wb_verbosity()
}
\value{
Logical indicating whether verbose output is enabled.
}
\description{
Get the workbench verbosity setting
}
\examples{
get_wb_verbosity()

}
