% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_setup.R
\name{have_wb}
\alias{have_wb}
\title{Check if Connectome Workbench is available}
\usage{
have_wb()
}
\value{
\code{TRUE} if \code{wb_command} is found and executable.
}
\description{
Check if Connectome Workbench is available
}
\examples{
have_wb()

}
