% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_setup.R
\name{set_wb_path}
\alias{set_wb_path}
\title{Set the path to wb_command}
\usage{
set_wb_path(path, check = TRUE)
}
\arguments{
\item{path}{Path to the \code{wb_command} executable.}

\item{check}{If \code{TRUE}, verify the path exists before setting.}
}
\value{
The previous value of the option, invisibly.
}
\description{
Set the path to wb_command
}
\examples{
\dontrun{
set_wb_path("/opt/workbench/bin/wb_command")
}

}
