% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_curvature.R
\name{surface_curvature}
\alias{surface_curvature}
\title{Compute surface curvature}
\usage{
surface_curvature(
  surface_in,
  mean_out = NULL,
  gauss_out = NULL,
  verbose = get_wb_verbosity()
)
}
\arguments{
\item{surface_in}{Path to the input surface file.}

\item{mean_out}{Path for the output mean curvature metric file.}

\item{gauss_out}{Path for the output Gaussian curvature metric file.}

\item{verbose}{Logical; print command output?}
}
\value{
The result of the underlying \code{\link[=wb_cmd]{wb_cmd()}} call.
}
\description{
Wraps \code{wb_command -surface-curvature} to compute mean and/or
Gaussian curvature of a surface.
}
\examples{
\dontrun{
surface_curvature(
  "lh.midthickness.surf.gii",
  mean_out = "lh.mean_curv.func.gii"
)
}

}
