% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_information.R
\name{surface_information}
\alias{surface_information}
\title{Display surface information}
\usage{
surface_information(surface_in, verbose = get_wb_verbosity())
}
\arguments{
\item{surface_in}{Path to the input surface file.}

\item{verbose}{Logical; print command output?}
}
\value{
Character vector of information lines, invisibly.
}
\description{
Wraps \code{wb_command -surface-information} to print summary information
about a surface file (vertices, triangles, bounding box, spacing).
}
\examples{
\dontrun{
surface_information("lh.midthickness.surf.gii")
}

}
