% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_vertex_areas.R
\name{surface_vertex_areas}
\alias{surface_vertex_areas}
\title{Compute vertex areas of a surface}
\usage{
surface_vertex_areas(surface_in, metric_out, verbose = get_wb_verbosity())
}
\arguments{
\item{surface_in}{Path to the input surface file.}

\item{metric_out}{Path for the output metric file.}

\item{verbose}{Logical; print command output?}
}
\value{
The result of the underlying \code{\link[=wb_cmd]{wb_cmd()}} call.
}
\description{
Wraps \code{wb_command -surface-vertex-areas} to measure the surface area
(in mm^2) that each vertex is responsible for.
}
\examples{
\dontrun{
surface_vertex_areas(
  "lh.midthickness.surf.gii",
  "lh.vertex_areas.func.gii"
)
}

}
\section{Connectome Workbench Help}{
\Sexpr[results=rd,stage=render]{wbcmd::wb_help_rd("-surface-vertex-areas")}
}

