% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_sitrep.R
\name{wb_sitrep}
\alias{wb_sitrep}
\title{Situation report for Connectome Workbench}
\usage{
wb_sitrep()
}
\value{
A list of diagnostic information, invisibly.
}
\description{
Prints a diagnostic summary of the workbench environment, including
path detection, version, and system information.
}
\examples{
wb_sitrep()

}
