% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_setup.R
\name{wb_version}
\alias{wb_version}
\title{Get Connectome Workbench version}
\usage{
wb_version()
}
\value{
Character string with the workbench version, or
an empty string if unavailable.
}
\description{
Get Connectome Workbench version
}
\examples{
\dontrun{
wb_version()
}

}
